requirements.txt
# Battery R&D Pipeline - Research Agent Dependencies

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# arXiv API
arxiv>=2.0.0

# RSS feeds
feedparser>=6.0.0

# GitHub API
requests>=2.31.0

# PDF processing
PyMuPDF>=1.23.0

# ML frameworks (for method implementations)
torch>=2.0.0
scikit-learn>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.0
pathlib>=1.0.0
